refactor uploaders
