+++ app/uploaders/document_file_uploader.rb
@@ -1,18 +1,12 @@
 # frozen_string_literal: true
 
-class DocumentFileUploader < CarrierWave::Uploader::Base
-  storage :file
-
+class DocumentFileUploader < ApplicationUploader
   def store_dir
     "uploads/operator_document_file/#{mounted_as}/#{model.id}"
   end
 
   def extension_whitelist
     %w(pdf doc docx txt csv xml jpg jpeg png exif tiff bmp)
-  end
-
-  def exists?
-    file.present?
   end
 
   def filename
@@ -26,13 +20,5 @@ class DocumentFileUploader < CarrierWave::Uploader::Base
     ].compact.join('-')
 
     filename + File.extname(super)
-  end
-
-  def original_filename
-    if file.present?
-      file.filename
-    else
-      super
-    end
   end
 end
