+++ app/uploaders/document_uploader.rb
@@ -1,17 +1,11 @@
 # frozen_string_literal: true
 
-class DocumentUploader < CarrierWave::Uploader::Base
-  storage :file
-
+class DocumentUploader < ApplicationUploader
   def store_dir
     "uploads/documents/#{model.id}"
   end
 
   def extension_whitelist
     %w(pdf doc docx txt csv xml jpg jpeg png exif tiff bmp)
-  end
-
-  def exists?
-    file.present?
   end
 end
