+++ app/uploaders/partner_logo_uploader.rb
@@ -1,18 +1,7 @@
 # frozen_string_literal: true
 
-class PartnerLogoUploader < CarrierWave::Uploader::Base
-  # Include RMagick or MiniMagick support:
-  # include CarrierWave::RMagick
+class PartnerLogoUploader < ApplicationUploader
   include CarrierWave::MiniMagick
-
-  # Choose what kind of storage to use for this uploader:
-  storage :file
-
-  # Override the directory where uploaded files will be stored.
-  # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-  end
 
   def extension_whitelist
     %w(jpg jpeg gif png svg tiff)
@@ -27,17 +16,5 @@ class PartnerLogoUploader < CarrierWave::Uploader::Base
   # This is just for active admin
   version :original do
     process resize_to_fill: [80, 80]
-  end
-
-  def exists?
-    file.present?
-  end
-
-  def original_filename
-    if file.present?
-      file.filename
-    else
-      super
-    end
   end
 end
