+++ app/uploaders/photo_uploader.rb
@@ -2,20 +2,8 @@
 
 require 'carrierwave/processing/mini_magick'
 
-class PhotoUploader < CarrierWave::Uploader::Base
-  # Include RMagick or MiniMagick support:
-  # include CarrierWave::RMagick
+class PhotoUploader < ApplicationUploader
   include CarrierWave::MiniMagick
-
-  # Choose what kind of storage to use for this uploader:
-  storage :file
-  # storage :fog
-
-  # Override the directory where uploaded files will be stored.
-  # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-  end
 
   def extension_whitelist
     %w(jpg jpeg gif png)
@@ -42,17 +30,5 @@ class PhotoUploader < CarrierWave::Uploader::Base
   # This is just for active admin
   version :original do
     process resize_to_fill: [80, 80]
-  end
-
-  def exists?
-    file.present?
-  end
-
-  def original_filename
-    if file.present?
-      file.filename
-    else
-      super
-    end
   end
 end
